+++ ext/trains.lua
@@ -3,10 +3,7 @@ local trainCoords = {}
 local storedTrains = {}
 
 Citizen.CreateThread(function()
-  -- Define the train models
   local trainModels = {"freight", "freightcar", "freightgrain", "freightcont1", "freightcont2","tankercar", "freighttrailer", "metrotrain"}
-
-  -- Define the train coordinates
   local trains = {
     {type=0, x=-498.4123, y=4304.3, z=88.40305},
     {type=0, x=2324.3, y=2670.7, z=44.45},
@@ -17,7 +14,6 @@ Citizen.CreateThread(function()
     {type=24, x=181.1, y=-1198.8, z=37.6},
   }
 
-  -- Load the train models
   for i= 1, 8 do
     RequestModel(GetHashKey(trainModels[i]))
     while not HasModelLoaded(GetHashKey(trainModels[i])) do
@@ -25,13 +21,11 @@ Citizen.CreateThread(function()
     end
   end
 
- -- Load the traindriver
   RequestModel(GetHashKey("s_m_m_lsmetro_01"))
   while not HasModelLoaded(GetHashKey("s_m_m_lsmetro_01")) do
     Citizen.Wait(1)
   end
 
-  -- Start the trains
   for i=1, 7, 1 do
     local trainCoords = trains[i]
     local thisTrain = CreateMissionTrain(trainCoords.type, trainCoords.x, trainCoords.y, trainCoords.z, true)